feat: handle gateway schema change events which occur in Slack channel
This is the initial commit of the socketmode Slack application for Team
Koko. Currently only processing of Slack message events has been
implemented.

This feature processes bot messages occuring in
gateway-schema-change-feed Slack channel. The gateway schema change
events rely on being able to parse a single attachment instance that
must contain a 'ref' and 'commit' field to extract the pull request
and org/repo in which the schema change event occurred in.